[Ionut] - W.I.P. Rework GUI element positioning and scaling - Update ImGUI and IMGUI_ADDONS
git-svn-id: https://xp-dev.com/svn/Divide-Framework/trunk@979 e89c6373-eeba-4702-abef-d26284fe765c

Former-commit-id: 088e052f4abae97c590cb580587078159e6977d3 [formerly bf041e490129d37617bc6c54c2ae3b111820ab1b] [formerly 2e8c2d195f241ecf764c4edfc0e21131b5981f2e [formerly 1c221a7d67835d3ecef3ef8498dcbc3c25f96280]]
Former-commit-id: 390bf09145c66ff1fef9d7e3129944cbc36c7f3d [formerly 36cb2778c69d259ecedab645306b6e183365df01]
Former-commit-id: eadeee82d82a6245caa3a1e8d83239767583d15a

+++ Source Code/Libs/imgui/addons/imguilistview/imguilistview.h
@@ -445,7 +445,7 @@ private:
         if (hasPrefix) {
             snprintf(&buf[bufid], pbufsz,"%s",prefix);
             bufid = strlen(buf);
-            pbufsz-= (int)bufid;
+            pbufsz-=bufid;
         }
 
         // value:
@@ -539,7 +539,7 @@ private:
         // suffix:
         if (hasSuffix)  {
             bufid = strlen(buf);
-            pbufsz-= (int)bufid;
+            pbufsz-=bufid;
             snprintf(&buf[bufid], pbufsz,"%s",suffix);
         }
